+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # TelegramBotAPI changelog
 
+## 18.2.2
+
+* `BehaviourBuilder`:
+  * Add `CommonMessageFilterExcludeCommand` to filter commands in messages
+  * Add `minus` operation for `SimpleFilter`s
+
 ## 18.2.1
 
 * `Version`: